requirements.txt
numpy
pillow
matplotlib
einops
scipy
constriction
jupyter